requirements.txt
# Project Dependencies
torch
torchvision
transformers
tensorflow
scikit-learn
pandas
numpy
matplotlib
seaborn
Pillow
